+++ app/src/main/java/com/don/prefaceapp/BookishApplication.kt
@@ -1,4 +1,4 @@
-package com.don.preface
+package com.don.prefaceapp
 
 import android.app.Application
 import android.util.Log
